Handle unsupported types without failing
We need to cope with:

* untyped properties (which default to null)
* Union types
* Intersection types

We can't guess anything for the latter two, as there's no reliable way
to guess which of the supported types might be appropriate. So we just
need the guessers to skip over them - previously they were failing due
to the ReflectionUnionType / ReflectionIntersectionType not supporting
getName() / isBuiltIn().